+++ app/gui/src/project-view/components/GraphEditor/GraphNode.vue
@@ -13,7 +13,6 @@ import GraphNodeMessage, {
   iconForMessageType,
   type MessageType,
 } from '@/components/GraphEditor/GraphNodeMessage.vue'
-import GraphNodeOutputPorts from '@/components/GraphEditor/GraphNodeOutputPorts.vue'
 import GraphVisualization from '@/components/GraphEditor/GraphVisualization.vue'
 import type { NodeCreationOptions } from '@/components/GraphEditor/nodeCreation'
 import PointFloatingMenu from '@/components/PointFloatingMenu.vue'
@@ -33,6 +32,7 @@ import { useNodeExecution } from '@/stores/project/nodeExecution'
 import { Ast } from '@/util/ast'
 import type { AstId } from '@/util/ast/abstract'
 import { prefixes } from '@/util/ast/node'
+import { onWindowBlur } from '@/util/autoBlur'
 import type { Opt } from '@/util/data/opt'
 import { Rect } from '@/util/data/rect'
 import { Vec2 } from '@/util/data/vec2'
@@ -67,7 +67,6 @@ const emit = defineEmits<{
   toggleDocPanel: []
   'update:edited': [cursorPosition: number]
   'update:rect': [rect: Rect]
-  'update:hoverAnim': [progress: number]
   'update:visualizationId': [id: Opt<VisualizationIdentifier>]
   'update:visualizationRect': [rect: Rect | undefined]
   'update:visualizationEnabled': [enabled: boolean]
@@ -158,7 +157,7 @@ const visibleMessage = computed(
     availableMessage.value,
 )
 
-const nodeHovered = ref(false)
+const nodeHovered = computed(() => graph.nodeHovered.get(nodeId.value) ?? false)
 
 const isOnlyOneSelected = computed(
   () =>
@@ -212,7 +211,6 @@ watch(menuVisible, (visible) => {
 
 function setSoleSelected() {
   nodeSelection?.setSelection(new Set([nodeId.value]))
-  graph.db.moveNodeToTop(nodeId.value)
 }
 
 function ensureSelected() {
@@ -221,7 +219,7 @@ function ensureSelected() {
   }
 }
 
-const outputHovered = ref(false)
+const outputHovered = computed(() => (graph.nodeOutputHoverAnimations.get(nodeId.value) ?? 0) !== 0)
 const keyboard = injectKeyboard()
 
 const visualizationWidth = computed(() => props.node.vis?.width ?? null)
@@ -232,14 +230,28 @@ const isVisualizationEnabled = computed({
     emit('update:visualizationEnabled', enabled)
   },
 })
+const visualizationHovered = ref(false)
+
 const isVisualizationPreviewed = computed(
-  () => keyboard.mod && outputHovered.value && !isVisualizationEnabled.value,
+  () =>
+    keyboard.mod &&
+    (outputHovered.value || visualizationHovered.value || nodeHovered.value) &&
+    !isVisualizationEnabled.value,
 )
 const isVisualizationVisible = computed(
   () => isVisualizationEnabled.value || isVisualizationPreviewed.value,
 )
+watch(isVisualizationVisible, (val) => {
+  // When visualization is being hidden, we don’t receive `pointerleave` event for some reason.
+  // So we need to set `visualizationHovered` to `false` manually.
+  if (!val) {
+    visualizationHovered.value = false
+  }
+})
 watch(isVisualizationPreviewed, (newVal, oldVal) => {
-  if (newVal && !oldVal) {
+  if (!newVal) {
+    graph.setNodeHovered(nodeId.value, false)
+  } else if (newVal && !oldVal) {
     graph.db.moveNodeToTop(nodeId.value)
   }
 })
@@ -453,6 +465,10 @@ const { editingComment } = provideComponentButtons(
 )
 
 const showMenuAt = ref<{ x: number; y: number }>()
+onWindowBlur(() => {
+  graph.setNodeHovered(nodeId.value, false)
+  updateNodeHover(undefined)
+})
 </script>
 
 <template>
@@ -464,8 +480,8 @@ const showMenuAt = ref<{ x: number; y: number }>()
     :class="nodeClass"
     :data-node-id="nodeId"
     @pointerdown.stop
-    @pointerenter="((nodeHovered = true), updateNodeHover($event))"
-    @pointerleave="((nodeHovered = false), updateNodeHover(undefined))"
+    @pointerenter="(graph.setNodeHovered(nodeId, true), updateNodeHover($event))"
+    @pointerleave="(graph.setNodeHovered(nodeId, false), updateNodeHover(undefined))"
     @pointermove="updateNodeHover"
   >
     <div class="binding" v-text="node.pattern?.code()" />
@@ -498,6 +514,8 @@ const showMenuAt = ref<{ x: number; y: number }>()
       :isPreview="isVisualizationPreviewed"
       :isFullscreenAllowed="true"
       :isResizable="true"
+      @pointerenter="visualizationHovered = true"
+      @pointerleave="visualizationHovered = false"
       @update:rect="updateVisualizationRect"
       @update:id="emit('update:visualizationId', $event)"
       @update:enabled="emit('update:visualizationEnabled', $event)"
@@ -544,21 +562,7 @@ const showMenuAt = ref<{ x: number; y: number }>()
       :message="visibleMessage.text"
       :type="visibleMessage.type"
     />
-    <svg class="bgPaths">
-      <rect class="bgFill" />
-      <GraphNodeOutputPorts
-        v-if="props.node.type !== 'output'"
-        :nodeId="nodeId"
-        :forceVisible="nodeHovered"
-        @newNodeClick="
-          (setSoleSelected(), emit('createNodes', [{ commit: false, content: undefined }]))
-        "
-        @portClick="(...args) => emit('outputPortClick', ...args)"
-        @portDoubleClick="(...args) => emit('outputPortDoubleClick', ...args)"
-        @update:hoverAnim="emit('update:hoverAnim', $event)"
-        @update:nodeHovered="outputHovered = $event"
-      />
-    </svg>
+    <div class="nodeBackground"></div>
   </div>
   <PointFloatingMenu v-if="showMenuAt" :point="showMenuAt" @close="showMenuAt = undefined">
     <ComponentContextMenu @close="showMenuAt = undefined" />
@@ -566,31 +570,22 @@ const showMenuAt = ref<{ x: number; y: number }>()
 </template>
 
 <style scoped>
-.bgPaths {
-  width: 100%;
-  height: 100%;
-  position: absolute;
-  overflow: visible;
-  top: 0;
-  left: 0;
-  display: flex;
-  --output-port-transform: translateY(var(--viz-below-node));
-}
-
-.bgFill {
-  width: var(--node-size-x);
-  height: var(--node-size-y);
-  rx: var(--node-border-radius);
-
-  fill: var(--color-node-background);
-  transition: fill 0.2s ease;
-}
-
 .GraphNode {
   position: absolute;
   border-radius: var(--node-border-radius);
   transition: box-shadow 0.2s ease-in-out;
   box-sizing: border-box;
+}
+
+.nodeBackground {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  border-radius: var(--node-border-radius);
+  background-color: var(--color-node-background);
+  transition: background-color 0.2s ease;
 }
 
 .content {